+++ GRDB/Core/DatabaseQueue.swift
@@ -7,19 +7,6 @@ import UIKit
 /// A DatabaseQueue serializes access to an SQLite database.
 public final class DatabaseQueue {
     
-    // MARK: - Configuration
-    
-    /// The database configuration
-    public var configuration: Configuration {
-        return serializedDatabase.configuration
-    }
-    
-    /// The path to the database.
-    ///
-    /// It is nil for in-memory databases.
-    public let path: String!
-    
-    
     // MARK: - Initializers
     
     /// Opens the SQLite database at path *path*.
@@ -33,7 +20,6 @@ public final class DatabaseQueue {
     ///     - configuration: A configuration.
     /// - throws: A DatabaseError whenever an SQLite error occurs.
     public init(path: String, configuration: Configuration = Configuration()) throws {
-        self.path = path
         store = try DatabaseStore(path: path, attributes: configuration.fileAttributes)
         serializedDatabase = try SerializedDatabase(
             path: path,
@@ -49,7 +35,6 @@ public final class DatabaseQueue {
     ///
     /// - parameter configuration: A configuration.
     public init(configuration: Configuration = Configuration()) {
-        path = nil
         store = nil
         serializedDatabase = try! SerializedDatabase(
             path: ":memory:",
@@ -66,6 +51,21 @@ public final class DatabaseQueue {
         NSNotificationCenter.defaultCenter().removeObserver(self)
     }
     #endif
+    
+    
+    // MARK: - Configuration
+    
+    /// The database configuration
+    public var configuration: Configuration {
+        return serializedDatabase.configuration
+    }
+    
+    /// The path to the database.
+    ///
+    /// It is nil for in-memory databases.
+    public var path: String! {
+        return store?.path
+    }
     
     
     // MARK: - Database access